feat(id): 新增清理函数 refactor(id): 重构异步刷新缓存机制

+++ team4u-id/src/main/java/org/team4u/id/infrastructure/seq/sp/StepSequenceProvider.java
@@ -16,7 +16,7 @@ public interface StepSequenceProvider extends SequenceProvider {
         /**
          * 步进
          */
-        private Long step = 1L;
+        private Integer step = 1;
         /**
          * 最大值
          */
